@@ -4,35 +4,38 @@ description: TRYHARD app showcase
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.18.2 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
 
+  assets_audio_player: ^3.0.6
   email_validator: ^2.1.17
   equatable: ^2.0.5
-  json_serializable: ^6.5.1
-  fluttertoast: ^8.0.9
-  flutter_bloc: ^8.1.1
-  flutter_svg: ^1.0.0
-  freezed: ^2.1.0+1
-  freezed_annotation: ^2.1.0
+  json_serializable: ^6.6.1
+  fluttertoast: ^8.2.1
+  flutter_bloc: ^8.1.2
+  flutter_svg: ^2.0.4
+  freezed: ^2.3.2
+  freezed_annotation: ^2.2.0
   intl: ^0.17.0
-  path_provider: ^2.0.11
+  path_provider: ^2.0.14
   uuid: ^3.0.5
   given_when_then: ^0.2.0
-  google_sign_in: ^5.4.1
-  cloud_firestore: ^4.1.0
-  firebase_core: ^2.3.0
-  firebase_auth: ^4.1.3
-  firebase_storage: ^11.0.6
-  go_router: ^5.1.10
+  google_sign_in: ^6.1.0
+  cloud_firestore: ^4.5.0
+  firebase_core: ^2.9.0
+  firebase_auth: ^4.4.0
+  firebase_storage: ^11.1.0
+  go_router: ^6.5.5
   get_it: ^7.2.0
-  injectable: ^2.0.0
-  injectable_generator: ^2.0.0
-  hydrated_bloc: ^9.0.0
-  json_annotation: ^4.7.0
-  bloc_test: ^9.1.0
+  injectable: ^2.1.1
+  injectable_generator: ^2.1.5
+  hydrated_bloc: ^9.1.0
+  json_annotation: ^4.8.0
+  bloc_test: ^9.1.1
+  rive: ^0.10.4
+  golden_toolkit: ^0.15.0
 
 
 dev_dependencies:
