@@ -1,11 +1,28 @@
 part of 'auth_cubit.dart';
 
-@freezed
-class AuthState with _$AuthState {
-  factory AuthState.notAuthorized() = _AuthStateNotAuthorized;
+sealed class AuthState extends Equatable {}
 
-  factory AuthState.authorized(AuthUser user) = _AuthStateAuthorized;
+final class AuthNotAuthorizedState extends AuthState {
+  AuthNotAuthorizedState() : super();
 
-  factory AuthState.fromJson(Map<String, dynamic> json) =>
-      _$AuthStateFromJson(json);
+  @override
+  List<Object?> get props => [];
+}
+
+final class AuthErrorState extends AuthState {
+  AuthErrorState({required this.message}) : super();
+
+  final String message;
+
+  @override
+  List<Object?> get props => [message];
+}
+
+final class AuthAuthorizedState extends AuthState {
+  AuthAuthorizedState({required this.user}) : super();
+
+  final AuthUser user;
+
+  @override
+  List<Object?> get props => [user];
 }
