@@ -3,7 +3,7 @@ import 'package:tryhard_showcase/app/data/datasource/models/user_info_basic.dart
 
 abstract class UserRepository {
   Future<UserProfile> getUserProfile({
-    required String? userGuid,
+    required String userGuid,
   });
 
   Future<void> createUserProfile({
