@@ -1,56 +1,67 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tryhard_showcase/app/constants/keys.dart';
+import 'package:tryhard_showcase/app/constants/strings/user_guids.dart';
 import 'package:tryhard_showcase/app/data/datasource/models/user.dart';
+import 'package:tryhard_showcase/app/di/di.dart';
 import 'package:tryhard_showcase/app/ui/components/avatar.dart';
 import 'package:tryhard_showcase/app/ui/styles/colors.dart';
 import 'package:tryhard_showcase/app/utils/toast.dart';
+import 'package:tryhard_showcase/features/auth/data/auth_repository.dart';
 import 'package:tryhard_showcase/features/auth/domain/auth_cubit/auth_cubit.dart';
-import 'package:tryhard_showcase/features/profile/domain/profile_cubit.dart';
+import 'package:tryhard_showcase/features/profile/domain/cubit/profile_cubit.dart';
+import 'package:tryhard_showcase/features/profile/domain/interactor/user_interactor.dart';
 
 class ProfileOverviewScreen extends StatelessWidget {
-  const ProfileOverviewScreen({
-    Key? key,
-  }) : super(key: key);
+  const ProfileOverviewScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return SafeArea(
-      child: Scaffold(
-        key: kOverviewScreen,
-        body: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Center(
-            child: BlocConsumer<ProfileCubit, ProfileState>(
-              builder: (context, state) {
-                return state.when(
-                  initial: () => const _UserProfileInitial(),
-                  loading: () => const CircularProgressIndicator(),
-                  loaded: (userProfile) =>
-                      _UserProfileOverview(userProfile: userProfile),
-                  error: (userProfile, error) => userProfile == null
-                      ? const _UserProfileInitial()
-                      : _UserProfileOverview(userProfile: userProfile),
-                );
-              },
-              listener: (context, state) {
-                state.whenOrNull(
-                  error: (_, error) => showServiceToast(error),
-                );
-              },
-            ),
-          ),
+    return BlocProvider(
+      create: (_) => ProfileCubit(
+        userInteractor: sl<UserInteractor>(),
+      )..loadUserProfile(
+          userGuid: sl<AuthRepository>().getCurrentUserId(),
+        ),
+      child: const Padding(
+        padding: EdgeInsets.all(16.0),
+        child: Center(
+          key: kOverviewScreen,
+          child: ProfileContentView(),
         ),
       ),
     );
   }
 }
 
+class ProfileContentView extends StatelessWidget {
+  const ProfileContentView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocConsumer<ProfileCubit, ProfileState>(
+      builder: (context, state) {
+        return state.when(
+          initial: () => const _UserProfileInitial(),
+          loading: () => const CircularProgressIndicator(),
+          loaded: (userProfile) =>
+              _UserProfileOverview(userProfile: userProfile),
+          error: (userProfile, error) => userProfile == null
+              ? const _UserProfileInitial()
+              : _UserProfileOverview(userProfile: userProfile),
+        );
+      },
+      listener: (context, state) {
+        state.whenOrNull(
+          error: (_, error) => showServiceToast(error),
+        );
+      },
+    );
+  }
+}
+
 class _UserProfileOverview extends StatelessWidget {
-  const _UserProfileOverview({
-    Key? key,
-    required this.userProfile,
-  }) : super(key: key);
+  const _UserProfileOverview({required this.userProfile});
 
   final UserProfile userProfile;
 
@@ -87,7 +98,6 @@ class _UserProfileOverview extends StatelessWidget {
         TextButton(
           key: kLogoutButton,
           onPressed: () {
-            context.read<ProfileCubit>().logout();
             context.read<AuthCubit>().logout();
           },
           child: Text(
@@ -101,9 +111,7 @@ class _UserProfileOverview extends StatelessWidget {
 }
 
 class _UserProfileInitial extends StatelessWidget {
-  const _UserProfileInitial({
-    Key? key,
-  }) : super(key: key);
+  const _UserProfileInitial();
 
   final textStyle = const TextStyle(
     fontSize: 16,
@@ -134,11 +142,11 @@ class _UserProfileInitial extends StatelessWidget {
         ),
         _spacer32,
         TextButton(
-          onPressed: () => context.read<ProfileCubit>().loadUserProfile(context
-                  .read<AuthCubit>()
-                  .state
-                  .whenOrNull(authorized: (user) => user.userGuid) ??
-              "-1"),
+          onPressed: () {
+            context.read<ProfileCubit>().loadUserProfile(
+                  userGuid: _getUserGuid(context),
+                );
+          },
           child: Text(
             'Reload',
             style: textButtonStyle,
@@ -146,6 +154,15 @@ class _UserProfileInitial extends StatelessWidget {
         ),
       ],
     );
+  }
+
+  String _getUserGuid(BuildContext context) {
+    final state = context.read<AuthCubit>().state;
+
+    if (state is AuthAuthorizedState) {
+      return state.user.userGuid ?? UserGuids.unknownGuid;
+    }
+    return UserGuids.unknownGuid;
   }
 }
 
